requirements.txt
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-vision==3.7.3
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.65.1
idna==3.7
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
requests==2.32.3
rsa==4.9
urllib3==2.2.2
